@@ -1,38 +1,19 @@
 {
-  "name": "webstac-to-cal",
+  "name": "webstactocal2",
   "version": "0.1.0",
   "private": true,
+  "dependencies": {
+    "lodash": "^4.17.21",
+    "luxon": "^2.3.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-ga": "^3.3.0"
+  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "deploy": "gh-pages -d build"
-  },
-  "dependencies": {
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.3",
-    "@testing-library/user-event": "^12.6.0",
-    "@types/bootstrap": "^4.5.0",
-    "@types/gapi": "^0.0.39",
-    "@types/gapi.auth2": "^0.0.52",
-    "@types/gapi.calendar": "^3.0.4",
-    "@types/jest": "^26.0.20",
-    "@types/jquery": "^3.5.1",
-    "@types/lodash": "^4.14.167",
-    "@types/node": "^12.19.14",
-    "@types/react": "^16.14.2",
-    "@types/react-dom": "^16.9.10",
-    "gh-pages": "^3.1.0",
-    "jquery": "^3.5.1",
-    "lodash": "^4.17.20",
-    "moment": "^2.29.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-ga": "^3.3.0",
-    "react-scripts": "^4.0.1",
-    "typescript": "^4.1.3",
-    "web-vitals": "^0.2.4"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
@@ -40,11 +21,32 @@
       "react-app/jest"
     ]
   },
-  "homepage": "https://smhsu.github.io/webstacToCal/",
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/gapi": "^0.0.41",
+    "@types/gapi.auth2": "^0.0.55",
+    "@types/gapi.calendar": "^3.0.6",
+    "@types/jest": "^27.4.0",
+    "@types/luxon": "^2.0.9",
+    "@types/node": "^16.11.17",
+    "@types/react": "^17.0.38",
+    "@types/react-dom": "^17.0.11",
+    "react-scripts": "5.0.0",
+    "typescript": "^4.5.4",
+    "web-vitals": "^2.1.2"
+  }
 }
