@@ -1,14 +1,14 @@
-import * as ReactGA from "react-ga";
+import ReactGA from "react-ga";
 
 const TRACKING_ID = "UA-58192647-1";
 const IS_PRODUCTION = process.env.NODE_ENV === "production";
 
 /**
  * Wrapper for Google Analytics.  Handles initialization and only sends events if the environment is production.
- * 
+ *
  * @author Silas Hsu
  */
-class Analytics {
+export class Analytics {
     private static isInitialized = false;
 
     /**
@@ -17,13 +17,13 @@ class Analytics {
     constructor() {
         if (!Analytics.isInitialized && IS_PRODUCTION) {
             ReactGA.initialize(TRACKING_ID);
-            ReactGA.set({anonymizeIp: true});
+            ReactGA.set({ anonymizeIp: true });
         }
     }
 
     /**
      * Sends a page view event.
-     * 
+     *
      * @param {string} path - relative path of the page
      */
     sendPageView(path: string) {
@@ -41,5 +41,3 @@ class Analytics {
         }
     }
 }
-
-export default Analytics;
